requirements.txt
streamlit>=1.50.0
fastapi>=0.118.0
uvicorn>=0.37.0
sqlalchemy>=2.0.43
psycopg2-binary>=2.9.10
python-dotenv>=1.1.1
pytesseract>=0.3.13
pillow>=11.3.0
httpx>=0.28.1
beautifulsoup4>=4.14.2
biopython>=1.85
openai>=2.1.0
scipy>=1.16.2
scikit-learn>=1.7.2
scikit-image>=0.24.0
sentence-transformers>=5.1.1
torch>=2.8.0
opencv-python>=4.12.0
python-multipart>=0.0.20
validators>=0.35.0
python-jose[cryptography]>=3.3.0
numpy>=2.2.6
pandas>=2.3.3
requests>=2.32.5
playwright>=1.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
email-validator>=2.1.0.post1
google-cloud-vision>=3.10.2
google-auth>=2.41.1
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.184.0
firebase-admin>=6.5.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
